+++ src/AudioCapture.hpp
@@ -2,12 +2,16 @@
 
 #include <SDL2/SDL.h>
 
+#include <memory>
 #include <string>
 #include <vector>
 
 namespace md3 {
 
 class ProjectMEngine;
+#ifdef MILKDROP3_MACOS_ARM64
+class MacSystemAudioCapture;
+#endif
 
 struct AudioDeviceInfo {
     int index{-1};
@@ -25,6 +29,7 @@ public:
     void start(const std::string& selector = {});
     void stop();
     void cycle();
+    void select(int index);
 
     [[nodiscard]] std::string deviceName() const;
     [[nodiscard]] int deviceIndex() const noexcept;
@@ -40,7 +45,9 @@ private:
     SDL_AudioDeviceID deviceId_{0};
     int deviceIndex_{-1};
     std::string deviceName_{"Default capture device"};
+#ifdef MILKDROP3_MACOS_ARM64
+    std::unique_ptr<MacSystemAudioCapture> systemAudio_;
+#endif
 };
 
 } // namespace md3
-